Remove operator <->
